App Re-Work
The original app would just sum up the money that you saved on stuff.
I never really used this app. The new application will allow you to
create items that you'd like to save for and you can select what you
want to save for by contributing saving to it and it will show you the
progress. And then once you've saved enough you can by it and save
for a new item.